add application resources for all components
Providing ArgoConfig only works with the Helm kind without this patch.
This is a problem because we want to produce an Application for every
supported component kind when rendering the platform.

This patch threads the ArgoConfig struct described in the Quickstart
guide through every supported component kind.

Generated from schematic with holos 0.93.4-15-g63d00bf

```
holos generate platform guide
```