add bank-of-holos frontend
With this patch the bank of holos frontend starts and is accessible at
https://bank.holos.localhost

None of the other services are running, however, so login fails.  The
site loads at least.

The purpose of using the bank demo is to organize the services into
multiple projects owned by multiple teams to demonstrate how holos makes
the configuration safer, easier, and more consistent.

With this patch we can see how ConfigMaps can be managed in one place in
projects/bank-of-holos/config.cue and then projected into each of the
frontend, identity, and ledger namespaces.