environments: scope the bank of holos down to a stack (#1)
Previously, the bank of holos was a single collection of components with
a platform wide scope.  For example, namespaces for the frontend,
backend, and, secrets were defined at the root and managed by the
platform namespaces component.

This patch establishes an alternative model, scoping all resources for
the bank of holos to a single stack.  A stack is aspect oriented,
meaning it is oriented to an environment and an owner.  For the shared
environments (dev, test, stage, prod) the owners are teams.  The
security team, the backend dev team, and the frontend dev teams.

For the sandbox environment the owner of all resources is an individual.

The purpose of structuring the code this way is to provide an example of
how Holos helps formalize and abstract the entire stack of the bank's
services into a relatively simple set of attributes.  What environment
is the stack for, who owns the stack?

Namespaces are the foundation, subsequent patches will establish stack
scoped components for previous platform scoped components, such as:

 1. SecretStores
 2. HTTPRoutes

Essentially, any resource previously registered with one platform scoped
component needs to instead register with a stack component.