environments: generate login passwords in-cluster
This patch adds a Job to generate the login password Secret in-cluster.
The secret is generated in the namespace managed by the security team.
The script to generate it is a plain file now, we use the Kustomize
config map generator to mix it into the fully rendered output.

We need to disableNameSuffixHash because kustomize doesn't transform the
job to the correct name for some reason.