Implement Podinfo with per-environment Values
PROBLEM:

We've added code to the repo demonstrating a way to handle environments,
but we haven't yet tackled the problem of a component with imported
default values that has a different value on a per-environment basis.

SOLUTION:

Add a `podinfo` component that uses the correct environment namespace,
then update the configuration to ensure there are three replicas in
`prod` and a single replica in other environments.

OUTCOME:

A very basic example showing different Helm chart values on
a per-environment basis.