kargo: move frontend promotion configuration to bank-web project
Previously the promotion process was managed in the security project.
They should be managed in the nonprod-bank-web and prod-bank-web
projects for nonprod and prod environments.

With this patch the frontend is correctly organized by the tier of the
environment.

The stages all pull directly from the warehouse still.  Next step is to
orgranize the environments into a sequence.